External linking with ARPACK/LAPACK/BLAS works now. Note that you need to have a fully functional LAPACK/ARPACK library to build externally. Eg. even if the internal BLAS functions are used you need to have a LAPACK library *with* BLAS, since LAPACK links against BLAS. The same with ARPACK, you need to have LAPACK and BLAS to, as ARPACK links against them, even if you don't use external LAPACK and BLAS.
Patches applied:

 * csardi@szaffi2/igraph--main--0.5--patch-131
   External linking with ARPACK/LAPACK/BLAS works now.

+++ src/arpack.h
@@ -29,31 +29,31 @@
 typedef struct igraph_arpack_options_t {
   /* INPUT */
   char bmat[1];			/* I-standard problem, G-generalized */
-  int n; 			/* Dimension of the eigenproblem */
+  long int n; 			/* Dimension of the eigenproblem */
   char which[2];		/* LA, SA, LM, SM, BE */
-  int nev;                      /* Number of eigenvalues to be computed */
+  long int nev;                 /* Number of eigenvalues to be computed */
   igraph_real_t tol;		/* Stopping criterion */
-  int ncv;			/* Number of columns in V */
-  int ldv;			/* Leading dimension of V */
-  int ishift;			/* 0-reverse comm., 1-exact with tridiagonal */
-  int mxiter;                   /* Maximum number of update iterations to take */
-  int nb;			/* Block size on the recurrence, only 1 works */
-  int mode;			/* The kind of problem to be solved (1-5)
+  long int ncv;			/* Number of columns in V */
+  long int ldv;			/* Leading dimension of V */
+  long int ishift;		/* 0-reverse comm., 1-exact with tridiagonal */
+  long int mxiter;              /* Maximum number of update iterations to take */
+  long int nb;			/* Block size on the recurrence, only 1 works */
+  long int mode;		/* The kind of problem to be solved (1-5)
 				   1: A*x=l*x, A symmetric
 				   2: A*x=l*M*x, A symm. M pos. def.
 				   3: K*x = l*M*x, K symm., M pos. semidef.
 				   4: K*x = l*KG*x, K s. pos. semidef. KG s. indef.
 				   5: A*x = l*M*x, A symm., M symm. pos. semidef. */
-  int start;			/* 0: random, 1: use the supplied vector */
-  int lworkl;			/* Size of temporary storage, default is fine */
+  long int start;		/* 0: random, 1: use the supplied vector */
+  long int lworkl;		/* Size of temporary storage, default is fine */
   igraph_real_t sigma;          /* The shift for modes 3,4,5 */  
   /* OUTPUT */
-  int info;			/* What happened, see docs */
-  int ierr;			/* What happened  in the dseupd call */
-  int noiter;			/* The number of iterations taken */
-  int numop;			/* Number of OP*x operations */
-  int numopb;			/* Number of B*x operations if BMAT='G' */
-  int numreo;			/* Number of steps of re-orthogonalizations */
+  long int info;		/* What happened, see docs */
+  long int ierr;		/* What happened  in the dseupd call */
+  long int noiter;		/* The number of iterations taken */
+  long int numop;		/* Number of OP*x operations */
+  long int numopb;		/* Number of B*x operations if BMAT='G' */
+  long int numreo;		/* Number of steps of re-orthogonalizations */
 
   /* INTERNAL */
   igraph_vector_t *v;
@@ -62,15 +62,15 @@ typedef struct igraph_arpack_options_t {
   igraph_vector_t *d;
   igraph_vector_t *resid;
   igraph_vector_t *ax;
-  int *select;
-  int iparam[11];
-  int ipntr[11];
+  long int *select;
+  long int iparam[11];
+  long int ipntr[11];
 } igraph_arpack_options_t;
 
 void igraph_arpack_options_init(igraph_arpack_options_t *o);
 
 typedef int igraph_arpack_function_t(igraph_real_t *to, const igraph_real_t *from,
-				     int n, void *extra);
+				     long int n, void *extra);
 
 int igraph_arpack_rssolve(igraph_arpack_function_t *fun, void *extra,
 			  igraph_arpack_options_t *options,
